requirements.txt
asgiref==3.2.7
Django==3.0.7
django-filer==1.7.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
easy-thumbnails==2.7
jsonfield==3.1.0
Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
